% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{clean}
\alias{clean}
\title{Theme clean}
\usage{
clean(
  font_size = 14,
  font_color = "#222222",
  header_font_size = 15,
  header_font_color = "#222222",
  cell_padding = 6,
  centered = FALSE
)
}
\arguments{
\item{font_size}{Numeric value representing the size of the font within the table (in px).
Default is 14.}

\item{font_color}{Color of the font for the text within the table.
Default is #222222.}

\item{header_font_size}{Numeric value representing the size of the font within the table (in px).
Default is 15.}

\item{header_font_color}{Color of the font for the header text.
Default is #222222.}

\item{cell_padding}{Numeric value representing the padding size between cells (in px).
Default is 6.}

\item{centered}{Logical: vertically center the contents of the table.
Default is FALSE.}
}
\value{
an object of class theme that is applied to a reactable table.
}
\description{
Simple clean-look theme
}
\examples{
data <- iris[10:29, ]

## Standard clean theme
reactable(data,
          theme = clean())

## Additional options applied
reactable(data,
          theme = clean(font_size = 12, font_color = "grey", cell_padding = 3))

}
