% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_spark.R
\name{highlight_bars}
\alias{highlight_bars}
\title{Color of highlight used in `react_sparkbar`.}
\usage{
highlight_bars(
  first = "transparent",
  last = "transparent",
  min = "transparent",
  max = "transparent"
)
}
\arguments{
\item{first, last, min, max}{The colors of first, last, min, and max bars}
}
\value{
a function that provides colors for specific bars.
}
\description{
Color of highlight used in `react_sparkbar`.
}
