% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_max.R
\name{highlight_max}
\alias{highlight_max}
\title{Highlights the maximum value in a column}
\usage{
highlight_max(data, font_color = "black", highlighter = NULL)
}
\arguments{
\item{data}{Dataset containing at least one numeric column.}

\item{font_color}{color to assign to maximum value in a column.
Default color is black.}

\item{highlighter}{color to assign the background of a cell containing maximum value in a column.}
}
\description{
The `highlight_max()` function assigns a font color and/or background color to the maximum value in a column.
    It should be placed within the style argument in reactable::colDef.
}
\examples{
data <- MASS::road[11:17, ]

## By default, the maximum value is bold with a black font color
reactable(data, defaultColDef = colDef(
    style = highlight_max(data)))

## Assign a different font color
reactable(data, defaultColDef = colDef(
    style = highlight_max(data, font_color = "red")))

## Highlight the background of the cell for the maximum value in each column
reactable(data, defaultColDef = colDef(
    style = highlight_max(data, highlighter = "yellow")))
}
