% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_spark.R
\name{highlight_points}
\alias{highlight_points}
\title{Color of points used in `react_sparkline`.}
\usage{
highlight_points(
  all = "transparent",
  first = "transparent",
  last = "transparent",
  min = "transparent",
  max = "transparent"
)
}
\arguments{
\item{all, first, last, min, max}{The colors of all, first, last, min, and max points.}
}
\value{
a function that provides colors for specific points.
}
\description{
Color of points used in `react_sparkline`.
}
