% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_img.R
\name{insert_img}
\alias{insert_img}
\title{Add image from web to rows in a column}
\usage{
insert_img(height = "36", width = "36", label = NULL)
}
\arguments{
\item{height}{A value given for the height of the image in px.
Default height is 36px.}

\item{width}{A value given for the width of the image in px.
Default width is 36px.}
}
\value{
a function that renders an image
    to a column containing a valid web link.
}
\description{
The `insert_img()` function adds images obtained from the web to a column within reactable.
    It should be placed within the cell argument in reactable::colDef.
}
\examples{
## If no image links are in the original dataset, you need to assign them like so:
library(dplyr)
data <- iris \%>\%
 mutate(
   img = case_when(
     Species == "setosa" ~ "https://upload.wikimedia.org/wikipedia/commons/thumb/5/56/Kosaciec_szczecinkowaty_Iris_setosa.jpg/800px-Kosaciec_szczecinkowaty_Iris_setosa.jpg",
     Species == "versicolor" ~ "https://upload.wikimedia.org/wikipedia/commons/thumb/4/41/Iris_versicolor_3.jpg/1920px-Iris_versicolor_3.jpg",
     Species == "virginica" ~ "https://www.fs.fed.us/wildflowers/beauty/iris/Blue_Flag/images/iris_virginica/iris_virginica_virginica.jpg",
     TRUE ~ "NA"))

## Then use insert_img() to display images
reactable(data,
columns = list(
 img = colDef(cell = insert_img())))

## By default, images are given a size of 36px by 36px,
## but you can adjust the size using height and width:
reactable(data,
columns = list(
 img = colDef(cell = insert_img(height = "50", width = "45"))))

## Optionally assign a label to the image from another column
reactable(data,
columns = list(
 img = colDef(cell = insert_img(label = "Species"))))

}
