% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop.R
\name{lollipop}
\alias{lollipop}
\title{Add a horizontal lollipop to rows in a column}
\usage{
lollipop(
  data,
  circle_color = "#1e90ff",
  bar_color = "lightgrey",
  number_fmt = NULL,
  bright_values = TRUE
)
}
\arguments{
\item{data}{Dataset containing at least one numeric column.}

\item{circle_color}{A single color or vector of colors.
If more than one color is provided, the colors will be assigned from lowest value to highest value in the column.
Default color is "#1e90ff".}

\item{bar_color}{A single color or vector of colors.
If more than one color is provided, the colors will create a left-to-right gradient.
Default color is "lightgrey".}

\item{number_fmt}{Optionally format numbers using formats from the scales package.
Default is set to NULL.}

\item{bright_values}{Display values with a dark-colored background in white text.
Default is set to TRUE but can be turned off by setting to FALSE.}
}
\value{
a function that applies a lollipop
    to a column of numeric values.
}
\description{
The `lollipop()` function adds a lollipop (a circle on the end of a bar with a value inside) to each row of a column.
    The length of the bars are relative to the value of the row in relation to other values within the same column.
    It should be placed within the cell argument in reactable::colDef.
}
\examples{
library(reactable)
data <- MASS::Cars93[20:49, c("Make", "MPG.city", "MPG.highway")]

## By default, colors for the lollipop bar and circle are provided
reactable(
data,
defaultColDef = colDef(
cell = lollipop(data)))

## Change the color of the bar with 'bar_color' and the circle with 'circle_color'
reactable(
data,
columns = list(
MPG.city = colDef(
 cell = lollipop(data,
 bar_color = "black",
 circle_color = "red")),
MPG.highway = colDef(
 cell = lollipop(data,
 bar_color = "orange",
 circle_color = "black"))))

## Providing more than one color to 'bar_color' applies a color gradient
## and providing more than one color to 'circle_color'
## conditionally colors based on their relative values
reactable(
data,
columns = list(
MPG.city = colDef(
 cell = lollipop(data,
 bar_color = c("#440154FF","#31688EFF","#35B779FF","#FDE725FF"),
 circle_color = "black")),
MPG.highway = colDef(cell = lollipop(data,
 bar_color = "black",
 circle_color = c("#440154FF","#31688EFF","#35B779FF","#FDE725FF")))))

}
