% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_elements.R
\name{margin}
\alias{margin}
\title{Margin dimensions.}
\usage{
margin(t = 0, r = 0, b = 0, l = 0)
}
\arguments{
\item{t, r, b, l}{The dimensions of the top, right, bottom, and left margins.}
}
\value{
a function that provides margin dimensions.
}
\description{
Margin dimensions.
}
