% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactablefmtr-package.R
\docType{package}
\name{reactablefmtr-package}
\alias{reactablefmtr}
\alias{reactablefmtr-package}
\title{reactablefmtr: Easily Customize Interactive Tables Made with Reactable}
\description{
Enhance the styling of interactive reactable tables with easy-to-use
    and highly-customizable functions. Apply conditional formatting to cells with
    data bars, color scales, and icon sets. Utilize custom table themes inspired by
    popular websites and bootstrap themes. Increase the portability and reproducibility
    of reactable tables by embedding images from the web directly into cells.
    Save the final table output as a static image or interactive file
    (note this feature requires the {webshot2} package which can be downloaded from https://github.com/rstudio/webshot2).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kcuilla.github.io/reactablefmtr/}
  \item \url{https://github.com/kcuilla/reactablefmtr}
  \item Report bugs at \url{https://github.com/kcuilla/reactablefmtr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Cuilla \email{kyle.cuilla@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Greg Lin [contributor]
  \item June Choe [contributor]
}

}
\keyword{internal}
