% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactablefmtr-package.R
\docType{package}
\name{reactablefmtr-package}
\alias{reactablefmtr}
\alias{reactablefmtr-package}
\title{reactablefmtr: Simplify Formatting of Tables Made with 'Reactable'}
\description{
Streamlines the process of applying formats and styles to tables using 'reactable'.
    Style your 'reactable' table with conditional formatting, color scales, data bars, and icons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kcuilla.github.io/reactablefmtr/}
  \item \url{https://github.com/kcuilla/reactablefmtr}
  \item Report bugs at \url{https://github.com/kcuilla/reactablefmtr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Cuilla \email{kyle.cuilla@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Greg Lin [contributor]
  \item June Choe [contributor]
}

}
\keyword{internal}
