% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactablefmtr-package.R
\docType{package}
\name{reactablefmtr-package}
\alias{reactablefmtr}
\alias{reactablefmtr-package}
\title{reactablefmtr: Streamlined Table Styling and Formatting for Reactable}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides various features to streamline and enhance the styling of interactive reactable tables with easy-to-use and highly-customizable functions and themes. Apply conditional formatting to cells with data bars, color scales, color tiles, and icon sets. Utilize custom table themes inspired by popular websites such and bootstrap themes. Apply sparkline line & bar charts (note this feature requires the 'dataui' package which can be downloaded from <https://github.com/timelyportfolio/dataui>). Increase the portability and reproducibility of reactable tables by embedding images from the web directly into cells. Save the final table output as a static image or interactive file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kcuilla.github.io/reactablefmtr/}
  \item \url{https://github.com/kcuilla/reactablefmtr}
  \item Report bugs at \url{https://github.com/kcuilla/reactablefmtr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Cuilla \email{kyle.cuilla@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Greg Lin [contributor]
  \item June Choe [contributor]
  \item Kent Russell [contributor]
}

}
\keyword{internal}
